test(governance): 拆分 app server events 输入路由测试
