+++ src/features/app/hooks/useAppServerEvents.runtime-ended.test.tsx
@@ -44,6 +44,97 @@ async function mount(handlers: Handlers) {
 }
 
 describe("useAppServerEvents runtime ended routing", () => {
+  it("routes runtime ended payloads and tears down shared thread processing", async () => {
+    const handlers: Handlers = {
+      onRuntimeEnded: vi.fn(),
+      onTurnError: vi.fn(),
+    };
+    registerSharedSessionNativeBinding({
+      workspaceId: "ws-runtime-ended",
+      sharedThreadId: "shared-thread-1",
+      nativeThreadId: "native-thread-1",
+      engine: "codex",
+    });
+    const { root } = await mount(handlers);
+
+    act(() => {
+      listener?.({
+        workspace_id: "ws-runtime-ended",
+        message: {
+          method: "runtime/ended",
+          params: {
+            reasonCode: "process_exit",
+            message: "Managed runtime process exited unexpectedly.",
+            affectedThreadIds: ["native-thread-1"],
+            affectedTurnIds: ["turn-77"],
+            pendingRequestCount: 2,
+            hadActiveLease: true,
+          },
+        },
+      });
+    });
+
+    expect(handlers.onRuntimeEnded).toHaveBeenCalledWith("ws-runtime-ended", {
+      reasonCode: "process_exit",
+      message: "Managed runtime process exited unexpectedly.",
+      affectedThreadIds: ["native-thread-1"],
+      affectedTurnIds: ["turn-77"],
+      pendingRequestCount: 2,
+      hadActiveLease: true,
+    });
+    expect(handlers.onTurnError).toHaveBeenCalledWith(
+      "ws-runtime-ended",
+      "shared-thread-1",
+      "turn-77",
+      expect.objectContaining({
+        message: "[RUNTIME_ENDED] Managed runtime process exited unexpectedly.",
+        willRetry: false,
+        engine: "codex",
+      }),
+    );
+
+    await act(async () => {
+      root.unmount();
+    });
+  });
+
+  it("skips turn teardown for manual shutdown runtime ended events", async () => {
+    const handlers: Handlers = {
+      onRuntimeEnded: vi.fn(),
+      onTurnError: vi.fn(),
+    };
+    const { root } = await mount(handlers);
+
+    act(() => {
+      listener?.({
+        workspace_id: "ws-runtime-ended-manual",
+        message: {
+          method: "runtime/ended",
+          params: {
+            reasonCode: "manual_shutdown",
+            message: "Managed runtime stopped after manual shutdown.",
+            affectedThreadIds: ["thread-1"],
+            affectedTurnIds: ["turn-1"],
+            pendingRequestCount: 0,
+            hadActiveLease: false,
+          },
+        },
+      });
+    });
+
+    expect(handlers.onRuntimeEnded).toHaveBeenCalledWith(
+      "ws-runtime-ended-manual",
+      expect.objectContaining({
+        reasonCode: "manual_shutdown",
+      }),
+    );
+    expect(handlers.onTurnError).not.toHaveBeenCalled();
+
+    await act(async () => {
+      root.unmount();
+    });
+  });
+
   it("uses explicit runtime ended thread-turn mappings for multi-thread teardown", async () => {
     const handlers: Handlers = {
       onTurnError: vi.fn(),
