+++ src/features/app/hooks/useAppServerEvents.test.tsx
@@ -59,271 +59,6 @@ async function mount(handlers: Handlers, options?: HookOptions) {
 }
 
 describe("useAppServerEvents", () => {
-  it("normalizes request user input questions and options", async () => {
-    const handlers: Handlers = {
-      onRequestUserInput: vi.fn(),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-9",
-        message: {
-          method: "item/tool/requestUserInput",
-          id: 55,
-          params: {
-            threadId: "thread-9",
-            turnId: "turn-9",
-            itemId: "item-9",
-            questions: [
-              {
-                id: "",
-                header: "",
-                question: "",
-                options: [
-                  { label: "", description: "" },
-                  { label: "  ", description: " " },
-                ],
-              },
-              {
-                id: "q-1",
-                header: "",
-                question: "Choose",
-                options: [
-                  { label: "", description: "" },
-                  { label: "Yes", description: "" },
-                  { label: "", description: "No label" },
-                ],
-              },
-            ],
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRequestUserInput).toHaveBeenCalledWith({
-      workspace_id: "ws-9",
-      request_id: 55,
-      params: {
-        thread_id: "thread-9",
-        turn_id: "turn-9",
-        item_id: "item-9",
-        questions: [
-          {
-            id: "q-1",
-            header: "",
-            question: "Choose",
-            isOther: false,
-            isSecret: false,
-            options: [
-              { label: "Yes", description: "" },
-              { label: "", description: "No label" },
-            ],
-          },
-        ],
-      },
-    });
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("routes runtime ended payloads and tears down shared thread processing", async () => {
-    const handlers: Handlers = {
-      onRuntimeEnded: vi.fn(),
-      onTurnError: vi.fn(),
-    };
-    registerSharedSessionNativeBinding({
-      workspaceId: "ws-runtime-ended",
-      sharedThreadId: "shared-thread-1",
-      nativeThreadId: "native-thread-1",
-      engine: "codex",
-    });
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-runtime-ended",
-        message: {
-          method: "runtime/ended",
-          params: {
-            reasonCode: "process_exit",
-            message: "Managed runtime process exited unexpectedly.",
-            affectedThreadIds: ["native-thread-1"],
-            affectedTurnIds: ["turn-77"],
-            pendingRequestCount: 2,
-            hadActiveLease: true,
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRuntimeEnded).toHaveBeenCalledWith("ws-runtime-ended", {
-      reasonCode: "process_exit",
-      message: "Managed runtime process exited unexpectedly.",
-      affectedThreadIds: ["native-thread-1"],
-      affectedTurnIds: ["turn-77"],
-      pendingRequestCount: 2,
-      hadActiveLease: true,
-    });
-    expect(handlers.onTurnError).toHaveBeenCalledWith(
-      "ws-runtime-ended",
-      "shared-thread-1",
-      "turn-77",
-      expect.objectContaining({
-        message: "[RUNTIME_ENDED] Managed runtime process exited unexpectedly.",
-        willRetry: false,
-        engine: "codex",
-      }),
-    );
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("skips turn teardown for manual shutdown runtime ended events", async () => {
-    const handlers: Handlers = {
-      onRuntimeEnded: vi.fn(),
-      onTurnError: vi.fn(),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-runtime-ended-manual",
-        message: {
-          method: "runtime/ended",
-          params: {
-            reasonCode: "manual_shutdown",
-            message: "Managed runtime stopped after manual shutdown.",
-            affectedThreadIds: ["thread-1"],
-            affectedTurnIds: ["turn-1"],
-            pendingRequestCount: 0,
-            hadActiveLease: false,
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRuntimeEnded).toHaveBeenCalledWith(
-      "ws-runtime-ended-manual",
-      expect.objectContaining({
-        reasonCode: "manual_shutdown",
-      }),
-    );
-    expect(handlers.onTurnError).not.toHaveBeenCalled();
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("preserves multiSelect flag for request user input questions", async () => {
-    const handlers: Handlers = {
-      onRequestUserInput: vi.fn(),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-multi",
-        message: {
-          method: "item/tool/requestUserInput",
-          id: "req-multi-1",
-          params: {
-            threadId: "thread-multi",
-            turnId: "turn-multi",
-            itemId: "item-multi",
-            questions: [
-              {
-                id: "q-1",
-                header: "Focus",
-                question: "Choose multiple",
-                multiSelect: true,
-                options: [{ label: "A", description: "" }],
-              },
-            ],
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRequestUserInput).toHaveBeenCalledWith({
-      workspace_id: "ws-multi",
-      request_id: "req-multi-1",
-      params: {
-        thread_id: "thread-multi",
-        turn_id: "turn-multi",
-        item_id: "item-multi",
-        questions: [
-          {
-            id: "q-1",
-            header: "Focus",
-            question: "Choose multiple",
-            isOther: false,
-            isSecret: false,
-            multiSelect: true,
-            options: [{ label: "A", description: "" }],
-          },
-        ],
-      },
-    });
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("prefers params.request_id over transport-level message.id for requestUserInput", async () => {
-    const handlers: Handlers = {
-      onRequestUserInput: vi.fn(),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-req-id",
-        message: {
-          method: "item/tool/requestUserInput",
-          id: 999,
-          params: {
-            request_id: "ask-real-1",
-            threadId: "thread-real",
-            turnId: "turn-real",
-            itemId: "item-real",
-            questions: [{ id: "q-1", header: "", question: "Choose one" }],
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRequestUserInput).toHaveBeenCalledWith({
-      workspace_id: "ws-req-id",
-      request_id: "ask-real-1",
-      params: {
-        thread_id: "thread-real",
-        turn_id: "turn-real",
-        item_id: "item-real",
-        questions: [
-          {
-            id: "q-1",
-            header: "",
-            question: "Choose one",
-            isOther: false,
-            isSecret: false,
-            options: undefined,
-          },
-        ],
-      },
-    });
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
   it("falls back to active codex thread for reasoning events without threadId", async () => {
     const handlers: Handlers = {
       onAppServerEvent: vi.fn(),
@@ -645,192 +380,6 @@ describe("useAppServerEvents", () => {
       threadId: "claude:session-2",
       itemId: "item-2",
       delta: "alias-delta",
-    });
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("normalizes secret input field from snake_case", async () => {
-    const handlers: Handlers = {
-      onRequestUserInput: vi.fn(),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-secret",
-        message: {
-          method: "item/tool/requestUserInput",
-          id: 87,
-          params: {
-            thread_id: "thread-secret",
-            turn_id: "turn-secret",
-            item_id: "item-secret",
-            questions: [
-              {
-                id: "token",
-                header: "Credential",
-                question: "Paste token",
-                is_secret: true,
-              },
-            ],
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRequestUserInput).toHaveBeenCalledWith({
-      workspace_id: "ws-secret",
-      request_id: 87,
-      params: {
-        thread_id: "thread-secret",
-        turn_id: "turn-secret",
-        item_id: "item-secret",
-        questions: [
-          {
-            id: "token",
-            header: "Credential",
-            question: "Paste token",
-            isOther: false,
-            isSecret: true,
-            options: undefined,
-          },
-        ],
-      },
-    });
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("falls back to turn.threadId and active codex thread for user input request", async () => {
-    const handlers: Handlers = {
-      onRequestUserInput: vi.fn(),
-      getActiveCodexThreadId: vi.fn(() => "codex-active-thread"),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-nested",
-        message: {
-          method: "item/tool/requestUserInput",
-          id: "req-nested-1",
-          params: {
-            turn: {
-              id: "turn-nested-1",
-              threadId: "thread-from-turn",
-            },
-            questions: [{ id: "q1", header: "", question: "Proceed?" }],
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRequestUserInput).toHaveBeenLastCalledWith({
-      workspace_id: "ws-nested",
-      request_id: "req-nested-1",
-      params: {
-        thread_id: "thread-from-turn",
-        turn_id: "turn-nested-1",
-        item_id: "",
-        questions: [
-          {
-            id: "q1",
-            header: "",
-            question: "Proceed?",
-            isOther: false,
-            isSecret: false,
-            options: undefined,
-          },
-        ],
-      },
-    });
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-nested",
-        message: {
-          method: "item/tool/requestUserInput",
-          id: "req-nested-2",
-          params: {
-            turnId: "turn-no-thread",
-            questions: [{ id: "q2", header: "", question: "Continue?" }],
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRequestUserInput).toHaveBeenLastCalledWith({
-      workspace_id: "ws-nested",
-      request_id: "req-nested-2",
-      params: {
-        thread_id: "codex-active-thread",
-        turn_id: "turn-no-thread",
-        item_id: "",
-        questions: [
-          {
-            id: "q2",
-            header: "",
-            question: "Continue?",
-            isOther: false,
-            isSecret: false,
-            options: undefined,
-          },
-        ],
-      },
-    });
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("marks requestUserInput as completed when payload indicates completion", async () => {
-    const handlers: Handlers = {
-      onRequestUserInput: vi.fn(),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-completed",
-        message: {
-          method: "item/tool/requestUserInput",
-          id: "req-completed-1",
-          params: {
-            threadId: "thread-completed",
-            turnId: "turn-completed",
-            itemId: "item-completed",
-            completed: true,
-            questions: [{ id: "q1", header: "", question: "Done?" }],
-          },
-        },
-      });
-    });
-
-    expect(handlers.onRequestUserInput).toHaveBeenCalledWith({
-      workspace_id: "ws-completed",
-      request_id: "req-completed-1",
-      params: {
-        thread_id: "thread-completed",
-        turn_id: "turn-completed",
-        item_id: "item-completed",
-        completed: true,
-        questions: [
-          {
-            id: "q1",
-            header: "",
-            question: "Done?",
-            isOther: false,
-            isSecret: false,
-            options: undefined,
-          },
-        ],
-      },
     });
 
     await act(async () => {
@@ -1588,33 +1137,6 @@ describe("useAppServerEvents", () => {
       "ws-1",
       "thread-1",
       3,
-    );
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("routes thread/compacted even when turnId is missing", async () => {
-    const handlers: Handlers = {
-      onContextCompacted: vi.fn(),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-2",
-        message: {
-          method: "thread/compacted",
-          params: { threadId: "thread-2" },
-        },
-      });
-    });
-
-    expect(handlers.onContextCompacted).toHaveBeenCalledWith(
-      "ws-2",
-      "thread-2",
-      "",
     );
 
     await act(async () => {
@@ -2872,120 +2394,6 @@ describe("useAppServerEvents", () => {
         aggregatedOutput: "stdout-line-1\nstdout-line-2",
         output: "stdout-line-1\nstdout-line-2",
       }),
-    );
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("keeps token usage updates when normalized realtime adapters handle item/completed", async () => {
-    const handlers: Handlers = {
-      onThreadTokenUsageUpdated: vi.fn(),
-      onAgentMessageCompleted: vi.fn(),
-    };
-    const { root } = await mount(handlers, {
-      useNormalizedRealtimeAdapters: true,
-    });
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-1",
-        message: {
-          method: "item/completed",
-          params: {
-            threadId: "thread-1",
-            item: { type: "agentMessage", id: "item-1", text: "Done" },
-            usage: {
-              input_tokens: 10,
-              output_tokens: 5,
-              cached_input_tokens: 2,
-              model_context_window: 128000,
-            },
-          },
-        },
-      });
-    });
-
-    expect(handlers.onAgentMessageCompleted).toHaveBeenCalledWith({
-      workspaceId: "ws-1",
-      threadId: "thread-1",
-      itemId: "item-1",
-      text: "Done",
-    });
-    expect(handlers.onThreadTokenUsageUpdated).toHaveBeenCalledWith(
-      "ws-1",
-      "thread-1",
-      {
-        total: {
-          inputTokens: 10,
-          outputTokens: 5,
-          cachedInputTokens: 2,
-          totalTokens: 15,
-        },
-        last: {
-          inputTokens: 10,
-          outputTokens: 5,
-          cachedInputTokens: 2,
-          totalTokens: 15,
-        },
-        modelContextWindow: 128000,
-      },
-    );
-
-    await act(async () => {
-      root.unmount();
-    });
-  });
-
-  it("prefers token_count last snapshot while keeping total snapshot", async () => {
-    const handlers: Handlers = {
-      onThreadTokenUsageUpdated: vi.fn(),
-      getActiveCodexThreadId: vi.fn(() => "thread-codex-1"),
-    };
-    const { root } = await mount(handlers);
-
-    act(() => {
-      listener?.({
-        workspace_id: "ws-1",
-        message: {
-          method: "token_count",
-          params: {
-            info: {
-              total_token_usage: {
-                input_tokens: 180000,
-                cached_input_tokens: 0,
-                model_context_window: 200000,
-              },
-              last_token_usage: {
-                input_tokens: 20000,
-                cached_input_tokens: 0,
-                model_context_window: 200000,
-              },
-            },
-          },
-        },
-      });
-    });
-
-    expect(handlers.onThreadTokenUsageUpdated).toHaveBeenCalledWith(
-      "ws-1",
-      "thread-codex-1",
-      {
-        total: {
-          inputTokens: 180000,
-          outputTokens: 0,
-          cachedInputTokens: 0,
-          totalTokens: 180000,
-        },
-        last: {
-          inputTokens: 20000,
-          outputTokens: 0,
-          cachedInputTokens: 0,
-          totalTokens: 20000,
-        },
-        modelContextWindow: 200000,
-      },
     );
 
     await act(async () => {
