+++ src/features/app/hooks/useAppServerEvents.tokenUsage.test.tsx
@@ -11,9 +11,16 @@ vi.mock("../../../services/events", () => ({
 }));
 
 type Handlers = Parameters<typeof useAppServerEvents>[0];
+type HookOptions = Parameters<typeof useAppServerEvents>[1];
 
-function TestHarness({ handlers }: { handlers: Handlers }) {
-  useAppServerEvents(handlers);
+function TestHarness({
+  handlers,
+  options,
+}: {
+  handlers: Handlers;
+  options?: HookOptions;
+}) {
+  useAppServerEvents(handlers, options);
   return null;
 }
 
@@ -33,16 +40,75 @@ afterEach(() => {
   vi.clearAllMocks();
 });
 
-async function mount(handlers: Handlers) {
+async function mount(handlers: Handlers, options?: HookOptions) {
   const container = document.createElement("div");
   const root = createRoot(container);
   await act(async () => {
-    root.render(<TestHarness handlers={handlers} />);
+    root.render(<TestHarness handlers={handlers} options={options} />);
   });
   return { root };
 }
 
 describe("useAppServerEvents token usage", () => {
+  it("keeps token usage updates when normalized realtime adapters handle item/completed", async () => {
+    const handlers: Handlers = {
+      onThreadTokenUsageUpdated: vi.fn(),
+      onAgentMessageCompleted: vi.fn(),
+    };
+    const { root } = await mount(handlers, {
+      useNormalizedRealtimeAdapters: true,
+    });
+
+    act(() => {
+      listener?.({
+        workspace_id: "ws-1",
+        message: {
+          method: "item/completed",
+          params: {
+            threadId: "thread-1",
+            item: { type: "agentMessage", id: "item-1", text: "Done" },
+            usage: {
+              input_tokens: 10,
+              output_tokens: 5,
+              cached_input_tokens: 2,
+              model_context_window: 128000,
+            },
+          },
+        },
+      });
+    });
+
+    expect(handlers.onAgentMessageCompleted).toHaveBeenCalledWith({
+      workspaceId: "ws-1",
+      threadId: "thread-1",
+      itemId: "item-1",
+      text: "Done",
+    });
+    expect(handlers.onThreadTokenUsageUpdated).toHaveBeenCalledWith(
+      "ws-1",
+      "thread-1",
+      {
+        total: {
+          inputTokens: 10,
+          outputTokens: 5,
+          cachedInputTokens: 2,
+          totalTokens: 15,
+        },
+        last: {
+          inputTokens: 10,
+          outputTokens: 5,
+          cachedInputTokens: 2,
+          totalTokens: 15,
+        },
+        modelContextWindow: 128000,
+      },
+    );
+
+    await act(async () => {
+      root.unmount();
+    });
+  });
+
   it("keeps token_count last usage as zero when only total snapshot exists", async () => {
     const handlers: Handlers = {
       onThreadTokenUsageUpdated: vi.fn(),
@@ -134,6 +200,61 @@ describe("useAppServerEvents token usage", () => {
           outputTokens: 0,
           cachedInputTokens: 12,
           totalTokens: 0,
+        },
+        modelContextWindow: 200000,
+      },
+    );
+
+    await act(async () => {
+      root.unmount();
+    });
+  });
+
+  it("prefers token_count last snapshot while keeping total snapshot", async () => {
+    const handlers: Handlers = {
+      onThreadTokenUsageUpdated: vi.fn(),
+      getActiveCodexThreadId: vi.fn(() => "thread-codex-1"),
+    };
+    const { root } = await mount(handlers);
+
+    act(() => {
+      listener?.({
+        workspace_id: "ws-1",
+        message: {
+          method: "token_count",
+          params: {
+            info: {
+              total_token_usage: {
+                input_tokens: 180000,
+                cached_input_tokens: 0,
+                model_context_window: 200000,
+              },
+              last_token_usage: {
+                input_tokens: 20000,
+                cached_input_tokens: 0,
+                model_context_window: 200000,
+              },
+            },
+          },
+        },
+      });
+    });
+
+    expect(handlers.onThreadTokenUsageUpdated).toHaveBeenCalledWith(
+      "ws-1",
+      "thread-codex-1",
+      {
+        total: {
+          inputTokens: 180000,
+          outputTokens: 0,
+          cachedInputTokens: 0,
+          totalTokens: 180000,
+        },
+        last: {
+          inputTokens: 20000,
+          outputTokens: 0,
+          cachedInputTokens: 0,
+          totalTokens: 20000,
         },
         modelContextWindow: 200000,
       },
